Face Verification/requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
facenet-pytorch==2.6.0
filelock==3.13.1
fonttools==4.58.4
fsspec==2024.6.1
greenlet==3.2.3
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.11.0.86
packaging==25.0
paho-mqtt==2.1.0
pandas==2.3.0
pillow==10.2.0
psutil==7.0.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
scipy==1.16.0
setuptools==70.2.0
six==1.17.0
SQLAlchemy==2.0.41
sympy==1.13.3
torch==2.2.2
torchaudio==2.7.1+cpu
torchvision==0.17.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.159
ultralytics-thop==2.0.14
urllib3==2.5.0
